Update cos-auditd-logging.yaml
Setting to Workers > 1 causes intermittent crashes of the fluent-bit service. This is explained here: https://github.com/fluent/fluent-bit/issues/5081

There is no fix in place. Instead, the latest fluent-bit code defaults the Workers to 1. I was going crazy trying to figure out what was going on. I blindly used the example here and I don't want anyone to lose time to this issue. 

I chose to set it to 1 explicitly so that this example can be compatible with older versions of fluent-bit as well.