Clean up socket tracer and deprecation reporter workloads
Make the sleep interval configurable in case the 60 second default is
too aggressive. Controlling this via an environment variable will allow
Autopilot users to change this value without needing to install an
updated WorkloadAllowlist.

Fix the deprecation reporter container name which was incorrect.

Remove modules volume which wasn't required for the socket-tracer to
function.

Add clarifying comments on why the debugfs volume and a privileged
security context are required.

Signed-off-by: Chris Henzie <chrishenzie@google.com>