AllowlistSynchronizers for socket-tracer
Enables GKE Autopilot customers to identify which workloads could be
calling the CRI v1alpha2 API.

The added annotations on the workloads are necessary for matching the
WorkloadAllowlist.

Also included are explicit ephemeral storage limits for the
socket-tracer container. This container uses around 350MiB for
installing packages, however Autopilot sets a default limit of 100MiB.
When this is exceeded the Pod is evicted. Setting this to 500MiB to
provide some wiggle room and allow the Pods to run on Autopilot nodes.

Signed-off-by: Chris Henzie <chrishenzie@google.com>